Prevent "no" and "idk" options for `self_in_case`
Forces the user to either be, or be filing for one of the participants
in the case on the side they say they are. They can't say they aren't a
part of the case, and then add themselves to the case.

Adds a note describing what to do if they aren't in the case.